migrate to new test suite
