+++ precompiles/staking/method_get_shares_test.go
@@ -17,34 +17,34 @@ import (
 func Test_GetShares(t *testing.T) {
 	t.Run("should return stakes", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, sdkKeepers, mockEVM, mockVMContract := setup(t)
-		methodID := abi.Methods[GetSharesMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[GetSharesMethodName]
 		r := rand.New(rand.NewSource(42))
 		validator := sample.Validator(t, r)
-		sdkKeepers.StakingKeeper.SetValidator(ctx, validator)
+		s.sdkKeepers.StakingKeeper.SetValidator(s.ctx, validator)
 
 		staker := sample.Bech32AccAddress()
 		stakerEthAddr := common.BytesToAddress(staker.Bytes())
 		coins := sample.Coins()
-		err := sdkKeepers.BankKeeper.MintCoins(ctx, fungibletypes.ModuleName, sample.Coins())
+		err := s.sdkKeepers.BankKeeper.MintCoins(s.ctx, fungibletypes.ModuleName, sample.Coins())
 		require.NoError(t, err)
-		err = sdkKeepers.BankKeeper.SendCoinsFromModuleToAccount(ctx, fungibletypes.ModuleName, staker, coins)
+		err = s.sdkKeepers.BankKeeper.SendCoinsFromModuleToAccount(s.ctx, fungibletypes.ModuleName, staker, coins)
 		require.NoError(t, err)
 
 		stakerAddr := common.BytesToAddress(staker.Bytes())
 
 		stakeArgs := []interface{}{stakerEthAddr, validator.OperatorAddress, coins.AmountOf(config.BaseDenom).BigInt()}
 
-		stakeMethodID := abi.Methods[StakeMethodName]
+		stakeMethodID := s.contractABI.Methods[StakeMethodName]
 
 		// ACT
-		_, err = contract.Stake(ctx, mockEVM, &vm.Contract{CallerAddress: stakerAddr}, &stakeMethodID, stakeArgs)
+		_, err = s.contract.Stake(s.ctx, s.mockEVM, &vm.Contract{CallerAddress: stakerAddr}, &stakeMethodID, stakeArgs)
 		require.NoError(t, err)
 
 		// ASSERT
 		args := []interface{}{stakerEthAddr, validator.OperatorAddress}
-		mockVMContract.Input = packInputArgs(t, methodID, args...)
-		stakes, err := contract.Run(mockEVM, mockVMContract, false)
+		s.mockVMContract.Input = packInputArgs(t, methodID, args...)
+		stakes, err := s.contract.Run(s.mockEVM, s.mockVMContract, false)
 		require.NoError(t, err)
 
 		res, err := methodID.Outputs.Unpack(stakes)
@@ -58,14 +58,14 @@ func Test_GetShares(t *testing.T) {
 
 	t.Run("should fail if wrong args amount", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, _, _, _ := setup(t)
-		methodID := abi.Methods[GetSharesMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[GetSharesMethodName]
 		staker := sample.Bech32AccAddress()
 		stakerEthAddr := common.BytesToAddress(staker.Bytes())
 		args := []interface{}{stakerEthAddr}
 
 		// ACT
-		_, err := contract.GetShares(ctx, &methodID, args)
+		_, err := s.contract.GetShares(s.ctx, &methodID, args)
 
 		// ASSERT
 		require.Error(t, err)
@@ -73,14 +73,14 @@ func Test_GetShares(t *testing.T) {
 
 	t.Run("should fail if invalid staker arg", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, _, _, _ := setup(t)
-		methodID := abi.Methods[GetSharesMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[GetSharesMethodName]
 		r := rand.New(rand.NewSource(42))
 		validator := sample.Validator(t, r)
 		args := []interface{}{42, validator.OperatorAddress}
 
 		// ACT
-		_, err := contract.GetShares(ctx, &methodID, args)
+		_, err := s.contract.GetShares(s.ctx, &methodID, args)
 
 		// ASSERT
 		require.Error(t, err)
@@ -88,14 +88,14 @@ func Test_GetShares(t *testing.T) {
 
 	t.Run("should fail if invalid val address", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, _, _, _ := setup(t)
-		methodID := abi.Methods[GetSharesMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[GetSharesMethodName]
 		staker := sample.Bech32AccAddress()
 		stakerEthAddr := common.BytesToAddress(staker.Bytes())
 		args := []interface{}{stakerEthAddr, staker.String()}
 
 		// ACT
-		_, err := contract.GetShares(ctx, &methodID, args)
+		_, err := s.contract.GetShares(s.ctx, &methodID, args)
 
 		// ASSERT
 		require.Error(t, err)
@@ -103,14 +103,14 @@ func Test_GetShares(t *testing.T) {
 
 	t.Run("should fail if invalid val address format", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, _, _, _ := setup(t)
-		methodID := abi.Methods[GetSharesMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[GetSharesMethodName]
 		staker := sample.Bech32AccAddress()
 		stakerEthAddr := common.BytesToAddress(staker.Bytes())
 		args := []interface{}{stakerEthAddr, 42}
 
 		// ACT
-		_, err := contract.GetShares(ctx, &methodID, args)
+		_, err := s.contract.GetShares(s.ctx, &methodID, args)
 
 		// ASSERT
 		require.Error(t, err)
