+++ precompiles/staking/method_move_stake_test.go
@@ -16,23 +16,23 @@ func Test_MoveStake(t *testing.T) {
 	// Disabled until further notice, check https://github.com/zeta-chain/node/issues/3005.
 	t.Run("should fail with error disabled", func(t *testing.T) {
 		// ARRANGE
-		ctx, contract, abi, sdkKeepers, mockEVM, mockVMContract := setup(t)
-		methodID := abi.Methods[MoveStakeMethodName]
+		s := newTestSuite(t)
+		methodID := s.contractABI.Methods[MoveStakeMethodName]
 		r := rand.New(rand.NewSource(42))
 		validatorSrc := sample.Validator(t, r)
-		sdkKeepers.StakingKeeper.SetValidator(ctx, validatorSrc)
+		s.sdkKeepers.StakingKeeper.SetValidator(s.ctx, validatorSrc)
 		validatorDest := sample.Validator(t, r)
 
 		staker := sample.Bech32AccAddress()
 		stakerEthAddr := common.BytesToAddress(staker.Bytes())
 		coins := sample.Coins()
-		err := sdkKeepers.BankKeeper.MintCoins(ctx, fungibletypes.ModuleName, sample.Coins())
+		err := s.sdkKeepers.BankKeeper.MintCoins(s.ctx, fungibletypes.ModuleName, sample.Coins())
 		require.NoError(t, err)
-		err = sdkKeepers.BankKeeper.SendCoinsFromModuleToAccount(ctx, fungibletypes.ModuleName, staker, coins)
+		err = s.sdkKeepers.BankKeeper.SendCoinsFromModuleToAccount(s.ctx, fungibletypes.ModuleName, staker, coins)
 		require.NoError(t, err)
 
 		stakerAddr := common.BytesToAddress(staker.Bytes())
-		mockVMContract.CallerAddress = stakerAddr
+		s.mockVMContract.CallerAddress = stakerAddr
 
 		argsStake := []interface{}{
 			stakerEthAddr,
@@ -41,9 +41,9 @@ func Test_MoveStake(t *testing.T) {
 		}
 
 		// stake to validator src
-		stakeMethodID := abi.Methods[StakeMethodName]
-		mockVMContract.Input = packInputArgs(t, stakeMethodID, argsStake...)
-		_, err = contract.Run(mockEVM, mockVMContract, false)
+		stakeMethodID := s.contractABI.Methods[StakeMethodName]
+		s.mockVMContract.Input = packInputArgs(t, stakeMethodID, argsStake...)
+		_, err = s.contract.Run(s.mockEVM, s.mockVMContract, false)
 		require.Error(t, err)
 		require.ErrorIs(t, err, ptypes.ErrDisabledMethod{
 			Method: StakeMethodName,
@@ -55,10 +55,10 @@ func Test_MoveStake(t *testing.T) {
 			validatorDest.OperatorAddress,
 			coins.AmountOf(config.BaseDenom).BigInt(),
 		}
-		mockVMContract.Input = packInputArgs(t, methodID, argsMoveStake...)
+		s.mockVMContract.Input = packInputArgs(t, methodID, argsMoveStake...)
 
 		// ACT
-		_, err = contract.Run(mockEVM, mockVMContract, false)
+		_, err = s.contract.Run(s.mockEVM, s.mockVMContract, false)
 
 		// ASSERT
 		require.Error(t, err)
